Extract shared AttendanceFormTable and implement AttendanceEditPage
Extract the attendance form UI (learner table, mark-all row, bottom bar
with counts, unsaved changes guard, mark-all confirmation modal) into a
shared AttendanceFormTable component and useAttendanceForm composable.
Refactor AttendanceNewPage to use these shared pieces.

Implement AttendanceEditPage that loads an existing attendance session,
displays records in the shared form table, tracks changes against the
original state, and saves only modified records via the bulk_update
endpoint.

Add AttendanceRecordResource for the new API endpoint, new attendance
strings, and update useAttendance composable with record fetching and
bulk update methods.
