+++ packages/kolibri-common/apiResources/AttendanceRecordResource.js
@@ -1,0 +1,8 @@
+import { Resource } from 'kolibri/apiResource';
+
+export default new Resource({
+  name: 'attendancerecord',
+  bulkUpdate(data) {
+    return this.postListEndpoint('bulk_update', data);
+  },
+});
